+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## \[Unreleased\]
 
 ### Added
+- Admin API call `graft_records`.
 ### Changed
+- Upgrade to Holochain v0.2.1.
 ### Fixed
 ### Removed
 
@@ -14,7 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add `storage_info` to the admin websocket.
 - Add `network_info` to the app websocket.
 ### Changed
-- Upgrade to Holochain 0.2 release candidate ahead of the holochain 0.2 release.
+- **BREAKING CHANGE**: Upgrade to Holochain 0.2 release candidate ahead of the holochain 0.2 release.
 
 ## 2023-02-15: v0.3.1
 ### Changed